+++ src/main/java/hr/tvz/financije/services/models/AccountDto.java
@@ -3,7 +3,7 @@ package hr.tvz.financije.services.models;
 public record AccountDto(
         int id,
         String name,
-        long amount,
+        double amount,
         String type,
         String color,
         String source,
